feat: Updated gradle and android gradle plugin to install app.
The error is
Unable to determine application id: com.android.tools.idea.run.ApkProvisionException: No outputs for the main artifact of variant: debug

The solution was to update the Android Gradle plugin from the current version 2.2.3 to version 3.6.0 and Gradle to version 5.6.4.
Now the app crashes after taking a picture and confirming on my tablet, API 17.